Refactor `asNormal` to be static in the normal entities & fix circular dependency
