+++ backend/src/database/entities/base-entity.ts
@@ -1,8 +1,6 @@
 import { IsUUID } from 'class-validator';
 import type { NormalType } from 'database/normal-type';
-import type { ServiceProvider } from 'database/services/service-provider';
 import { UUID, uuidValidationOptions } from 'digital-fuesim-manv-shared';
-import type { EntityManager } from 'typeorm';
 import { PrimaryGeneratedColumn } from 'typeorm';
 
 export abstract class BaseEntity<
@@ -12,10 +10,4 @@ export abstract class BaseEntity<
     @PrimaryGeneratedColumn('uuid')
     @IsUUID(4, uuidValidationOptions)
     id!: UUID;
-
-    // TODO: Better name for the whole `normal`/`Normal` stuff
-    abstract asNormal(
-        services: ServiceProvider,
-        entityManager?: EntityManager
-    ): NormalEntity | Promise<NormalEntity>;
 }
