+++ backend/src/database/entities/exercise-wrapper.entity.ts
@@ -1,128 +1,61 @@
-// import { ExerciseWrapperEntityAll } from './all-entities';
+import { IsInt, IsJSON, IsString, MinLength, MaxLength } from 'class-validator';
+import type { ServiceProvider } from 'database/services/service-provider';
+import { ExerciseState } from 'digital-fuesim-manv-shared';
+import type { ExerciseWrapper } from 'exercise/exercise-wrapper';
+import { Entity, Column } from 'typeorm';
+import { BaseEntity } from './base-entity';
 
-// // eslint-disable-next-line @typescript-eslint/naming-convention
-// export const ExerciseWrapperEntity = ExerciseWrapperEntityAll;
-// // eslint-disable-next-line @typescript-eslint/no-redeclare
-// export type ExerciseWrapperEntity = ExerciseWrapperEntityAll;
+@Entity()
+export class ExerciseWrapperEntity extends BaseEntity<
+    ExerciseWrapperEntity,
+    ExerciseWrapper
+> {
+    @Column({
+        type: 'integer',
+        default: 0,
+    })
+    @IsInt()
+    tickCounter = 0;
 
-// // import { IsInt, IsJSON, IsString, MinLength, MaxLength } from 'class-validator';
-// // import { ExerciseState } from 'digital-fuesim-manv-shared';
-// // import type { EntityManager } from 'typeorm';
-// // import { Entity, IsNull, Column } from 'typeorm';
-// // import type { ServiceProvider } from '../services/service-provider';
-// // import type { ActionWrapper } from '../../exercise/action-wrapper';
-// // import { ExerciseWrapper } from '../../exercise/exercise-wrapper';
-// // import { BaseEntity } from './base-entity';
+    @Column({
+        type: 'json',
+    })
+    @IsJSON()
+    initialStateString!: string;
 
-// // @Entity()
-// // export class ExerciseWrapperEntity extends BaseEntity<
-// //     ExerciseWrapperEntity,
-// //     ExerciseWrapper
-// // > {
-// //     async asNormal(
-// //         services: ServiceProvider,
-// //         entityManager?: EntityManager
-// //     ): Promise<ExerciseWrapper> {
-// //         const operations = async (manager: EntityManager) => {
-// //             const actions = await services.actionWrapperService.findAll(
-// //                 {
-// //                     where: {
-// //                         emitter: {
-// //                             exercise: {
-// //                                 id: this.id,
-// //                             },
-// //                         },
-// //                     },
-// //                     order: {
-// //                         created: 'ASC',
-// //                     },
-// //                 },
-// //                 manager
-// //             );
-// //             const actionsInWrapper: ActionWrapper[] = [];
-// //             const normal = new ExerciseWrapper(
-// //                 this.participantId,
-// //                 this.trainerId,
-// //                 actionsInWrapper,
-// //                 services,
-// //                 JSON.parse(this.initialStateString) as ExerciseState,
-// //                 (
-// //                     await services.actionEmitterService.findOne(
-// //                         {
-// //                             where: {
-// //                                 exercise: { id: this.id },
-// //                                 emitterName: IsNull(),
-// //                             },
-// //                         },
-// //                         false,
-// //                         manager
-// //                     )
-// //                 )?.emitterId ?? undefined
-// //             );
-// //             actionsInWrapper.splice(
-// //                 0,
-// //                 0,
-// //                 ...(await Promise.all(
-// //                     actions.map(async (action) =>
-// //                         action.asNormal(services, manager, normal)
-// //                     )
-// //                 ))
-// //             );
-// //             normal.id = this.id;
-// //             normal.tickCounter = this.tickCounter;
-// //             return normal;
-// //         };
-// //         return entityManager
-// //             ? operations(entityManager)
-// //             : services.transaction(operations);
-// //     }
+    @Column({ type: 'char', length: 6 })
+    @IsString()
+    @MinLength(6)
+    @MaxLength(6)
+    participantId!: string;
 
-// //     @Column({
-// //         type: 'integer',
-// //         default: 0,
-// //     })
-// //     @IsInt()
-// //     tickCounter = 0;
+    @Column({ type: 'char', length: 8 })
+    @IsString()
+    @MinLength(8)
+    @MaxLength(8)
+    trainerId!: string;
 
-// //     @Column({
-// //         type: 'json',
-// //     })
-// //     @IsJSON()
-// //     initialStateString!: string;
+    /**
+     * Be very careful when using this. - Use {@link create} instead for most use cases.
+     * This constructor does not guarantee a valid entity.
+     */
+    // eslint-disable-next-line @typescript-eslint/no-useless-constructor
+    constructor() {
+        super();
+    }
 
-// //     @Column({ type: 'char', length: 6 })
-// //     @IsString()
-// //     @MinLength(6)
-// //     @MaxLength(6)
-// //     participantId!: string;
+    static async create(
+        participantId: string,
+        trainerId: string,
+        services: ServiceProvider,
+        initialState: ExerciseState = ExerciseState.create()
+    ): Promise<ExerciseWrapperEntity> {
+        const exercise = await services.exerciseWrapperService.create({
+            participantId,
+            trainerId,
+            initialStateString: JSON.stringify(initialState),
+        });
 
-// //     @Column({ type: 'char', length: 8 })
-// //     @IsString()
-// //     @MinLength(8)
-// //     @MaxLength(8)
-// //     trainerId!: string;
-
-// //     /**
-// //      * Be very careful when using this. - Use {@link create} instead for most use cases.
-// //      * This constructor does not guarantee a valid entity.
-// //      */
-// //     // eslint-disable-next-line @typescript-eslint/no-useless-constructor
-// //     constructor() {
-// //         super();
-// //     }
-
-// //     static async create(
-// //         participantId: string,
-// //         trainerId: string,
-// //         services: ServiceProvider,
-// //         initialState: ExerciseState = ExerciseState.create()
-// //     ): Promise<ExerciseWrapperEntity> {
-// //         const exercise = await services.exerciseWrapperService.create({
-// //             participantId,
-// //             trainerId,
-// //             initialStateString: JSON.stringify(initialState),
-// //         });
-
-// //         return exercise;
-// //     }
-// // }
+        return exercise;
+    }
+}
